requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1

# HTTP and parsing
requests==2.32.3

# Vector store
faiss-cpu==1.9.0.post1
numpy==2.1.3

# LLM
groq==0.13.1

# Embeddings
gravixlayer==0.0.50

# PDF processing
PyMuPDF==1.25.1
langchain-text-splitters==0.3.3

# Pydantic for validation
pydantic==2.10.3

# Hybrid retrieval
rank-bm25==0.2.2